+++ Makefile
@@ -20,7 +20,7 @@ coderd/database/generate: fmt/sql coderd/database/dump.sql $(wildcard coderd/dat
 .PHONY: coderd/database/generate
 
 apitypings/generate: site/src/api/types.ts
-	go run scripts/apitypings/main.go > site/src/api/types-generated.ts
+	go run scripts/apitypings/main.go > site/src/api/typesGenerated.ts
 	cd site && yarn run format:types
 .PHONY: apitypings/generate
