+++ site/src/AppRouter.tsx
@@ -10,9 +10,9 @@ import { HealthzPage } from "./pages/healthz"
 import { SignInPage } from "./pages/login"
 import { OrganizationsPage } from "./pages/orgs"
 import { PreferencesAccountPage } from "./pages/preferences/account"
-import { PreferencesLinkedAccountsPage } from "./pages/preferences/linked-accounts"
+import { PreferencesLinkedAccountsPage } from "./pages/preferences/linkedAccounts"
 import { PreferencesSecurityPage } from "./pages/preferences/security"
-import { PreferencesSSHKeysPage } from "./pages/preferences/ssh-keys"
+import { PreferencesSSHKeysPage } from "./pages/preferences/sshKeys"
 import { SettingsPage } from "./pages/settings"
 import { TemplatesPage } from "./pages/templates"
 import { TemplatePage } from "./pages/templates/[organization]/[template]"
