+++ site/src/components/Icons/index.test.tsx
@@ -1,6 +1,6 @@
 import SvgIcon from "@material-ui/core/SvgIcon"
 import React from "react"
-import { render } from "./../../test_helpers"
+import { render } from "../../testHelpers"
 import * as Icons from "./index"
 
 const getAllIcons = (): [string, typeof SvgIcon][] => {
