+++ site/src/components/Page/Footer.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from "@testing-library/react"
 import React from "react"
-import { MockBuildInfo, render } from "../../test_helpers"
+import { MockBuildInfo, render } from "../../testHelpers"
 import { Footer, Language } from "./Footer"
 
 describe("Footer", () => {
