+++ site/src/components/Workspace/Workspace.stories.tsx
@@ -1,6 +1,6 @@
 import { Story } from "@storybook/react"
 import React from "react"
-import { MockOrganization, MockTemplate, MockWorkspace } from "../../test_helpers"
+import { MockOrganization, MockTemplate, MockWorkspace } from "../../testHelpers"
 import { Workspace, WorkspaceProps } from "./Workspace"
 
 export default {
