+++ site/src/components/Workspace/Workspace.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from "@testing-library/react"
 import React from "react"
-import { MockOrganization, MockTemplate, MockWorkspace, render } from "../../test_helpers"
+import { MockOrganization, MockTemplate, MockWorkspace, render } from "../../testHelpers"
 import { Workspace } from "./Workspace"
 
 describe("Workspace", () => {
