+++ site/src/util/firstLetter.ts
